#19: Rearranged test for feature and method based
Related Work Items: #1, #19
